Write cleaned up dataset to excel format
