+++ pyproject.toml
@@ -8,6 +8,7 @@ authors = ["Andreas Brodersen <43907402+AndreasPB@users.noreply.github.com>"]
 python = "^3.9"
 ipykernel = "^6.13.0"
 pandas = "^1.4.2"
+openpyxl = "^3.0.9"
 
 [tool.poetry.dev-dependencies]
